assignment 2 GradeChecker.py modified

+++ PythonAndBash/GradeChecker.py
@@ -1,4 +1,4 @@
-score=float(input("Enter your score: "))
+score=int(input("Enter your score: "))
 grade=""
 if score >= 90:
     grade="A"
